for #18: bamliquidator_regions now honors bamliquidator logging arguments
also samtools stderr logging is suppressed when in quiet mode.  this could possibly
lead to useful information being lossed, but this seems unlikely since samtool error
messages are usually pretty generic like the following:

  [bam_parse_region] fail to determine the sequence name.

and this would always lead to a line like the following in log.txt:

  2014-08-07 23:44:39 WARNING skipping region 2 (bam file key 1 chr33 MM1S_CTCF_11890 27130338 -> 27133471 . 0) due to error: bam_parse_region failed with return code -1

I'm not sure the "fail to determine the sequence name" is terribly helpful anyway.