removed no longer needed stderr suppression
this was just added in because previously a bad region would cause samtools to write to stderr,
which could clog up output.  but now we check the region before calling samtools, so we
shouldn't run into this case anymore.  if samtools does error despite the check, bamliquidator halts,
and writing to stderr here is ok