bamliquidator_batch now compatible with latest bokeh version
bamliquidator_batch was originally developed for bokeh 0.4.  latest
version is 0.9 and has some breaking changes.  for a while we held off
upgrading due to some performance problems with bokeh 0.5, but latest
version has good performance and 0.4 is too old to keep trying to use.

resolves #26
fixes one of the errors in #49