Correct binsize calculation
Binsize calculation is based on the gff input file and thus needs to add 1 for the region size. Since this was missing
an offset error accumulated with the number of bins used and the end of a region was never been used for calculation.
E.g. A region with 60 bins would lack the last 60bp of a region due to the offset error accumulation.